Move basic.py to where ci-cd will pick it up from
