+++ .gitignore
@@ -80,8 +80,6 @@ __pycache__
 
 *sgx_linux_x64_sdk_2.9.101.2.bin
 
-integ-tests/entropy_config
-
 # mob prompt and build tmp
 .mob/
 .tmp/
